protocol: fix panic in entity::Id JSON schema generation
serde_json cannot represent u128::MAX as a JSON number without
arbitrary_precision, so schema_for! on any type embedding entity::Id
panicked at runtime. Use the f64 approximation for the schema bound;
the exact range is still enforced at deserialization.

+++ protocol/src/types/entity.rs
@@ -566,7 +566,9 @@ impl JsonSchema for Id {
                 {
                     "type": "integer",
                     "minimum": 0,
-                    "maximum": u128::MAX,
+                    // JSON numbers cannot hold u128, so the bound is approximate;
+                    // the exact range is enforced at deserialization.
+                    "maximum": u128::MAX as f64,
                     "description": "integer representation of the identifier, only in binary formats"
                 },
                 {
